+++ src/effects/shaders/transition/slide-shader.ts
